Модули адаптированы для работы через Ajax. Добавлена функция удаления файлов в смайликах. Прочие доработки.

+++ modules/smilies/admin.php
@@ -15,8 +15,39 @@ if(!$user->CheckAccess2('smilies', 'smilies')){
 $smilies_dir = 'uploads/smilies/';
 $mod = ADMIN_FILE.'?exe=smilies';
 
-function AdminSmilesGetAllSmiles( &$sid, $dir_name, $selected = '', $smilies = array() )
-{
+if(isset($_GET['a'])){
+	$action = $_GET['a'];
+}else{
+	$action = 'main';
+}
+
+TAddToolLink('Смайлики', 'main', 'smilies');
+TAddToolLink('Авто-добавление', 'auto', 'smilies&a=auto');
+TAddToolBox($action);
+switch($action){
+	case 'main': AdminSmilesMain();
+		break;
+	case 'addsmile': AdminSmilesAddSave();
+		break;
+	case 'delsmile': AdminSmilesDeleteSmile();
+		break;
+	case 'editsmile': AdminSmilesEditSmile();
+		break;
+	case 'seditsave': AdminSmilesEditSave();
+		break;
+	case 'changestatus': AdminBlocksChangeStatus();
+		break;
+	case 'auto': AdminSmiliesAutoAdd();
+		break;
+	case 'autosave': AdminSmiliesAutoSave();
+		break;
+	case 'delfile': AdminSmiliesDeleteFile();
+		break;
+	default:
+		AdminSmilesMain();
+}
+
+function AdminSmilesGetAllSmiles( &$sid, $dir_name, $selected = '', $smilies = array() ){
 	global $site, $smilies_dir;
 	static $i = -1;
 	static $sfiles = array();
@@ -51,57 +82,59 @@ function AdminSmilesGetAllSmiles( &$sid, $dir_name, $selected = '', $smilies = a
 	return $sfiles;
 }
 
-function AdminSmilesMain()
-{
+function AdminSmilesMain(){
 	global $db, $config, $site, $smilies_dir, $mod;
+	System::admin()->AddCenterBox('Смайлики');
 	$smilies = $db->Select('smilies', '');
 	$text = '<table cellspacing="0" cellpadding="0" class="cfgtable">';
-	$text .= '<tr><th>Изображение</th><th>Код</th><th>Описание</th><th>Имя файла</th><th>Включить</th><th>Функции</th></tr>';
-	while($row = $db->FetchRow()){
-		$sid = SafeDB($row['id'], 11, int);
-		if(!is_file($smilies_dir.$row['file'])){
-			$db->Delete('smilies', "`file`='".SafeEnv($row['file'], 255, str)."'");
-			continue;
-		}
-		switch($row['enabled']){
-			case "1":
-				$en = '<a href="'.$mod.'&a=changestatus&id='.$sid.'" title="Изменить статус"><font color="#008000">Да</font></a>';
-				break;
-			case "0":
-				$en = '<a href="'.$mod.'&a=changestatus&id='.$sid.'" title="Изменить статус"><font color="#FF0000">Нет</font></a>';
-				break;
-		}
+	$text .= '<tr><th>Изображение</th><th>Код</th><th>Описание</th><th>Имя файла</th><th>Показать</th><th>Функции</th></tr>';
+	if(System::database()->NumRows() > 0){
+		while($row = $db->FetchRow()){
+			$sid = SafeDB($row['id'], 11, int);
+			if(!is_file($smilies_dir.$row['file'])){
+				$db->Delete('smilies', "`file`='".SafeEnv($row['file'], 255, str)."'");
+				continue;
+			}
+			$en = System::admin()->SpeedStatus('Да', 'Нет', $mod.'&a=changestatus&id='.$sid, $row['enabled'] == '1');
 
-		$func = '';
-		$func .= SpeedButton('Редактировать', $mod.'&a=editsmile&id='.$sid, 'images/admin/edit.png');
-		$func .= SpeedButton('Удалить', $mod.'&a=delsmile&sid='.$sid, 'images/admin/delete.png');
+			$func = '';
+			$func .= System::admin()->SpeedButton('Редактировать', $mod.'&a=editsmile&id='.$sid, 'images/admin/edit.png');
+			$func .= System::admin()->SpeedConfirm('Удалить', $mod.'&a=delsmile&sid='.$sid, 'images/admin/delete.png', 'Удалить смайлик?');
 
-		$text .= "<tr><td><a href=\"$mod&a=editsmile&id=$sid\"><img src=\"$smilies_dir{$row['file']}\" /></a></td><td>{$row['code']}</td><td>{$row['desc']}</td><td>{$row['file']}</td><td>$en</td><td>$func</td></tr>";
+			$text .= "<tr>
+			<td>".System::admin()->Link('<img src="'.$smilies_dir.$row['file'].'">', $mod.'&a=editsmile&id='.$sid)."</td>
+			<td>{$row['code']}</td>
+			<td>{$row['desc']}</td>
+			<td>{$row['file']}</td>
+			<td>$en</td>
+			<td>$func</td>
+			</tr>";
+		}
+	}else{
+		$text .= '<tr><td colspan="6" style="text-align: left;">Нет смайликов.</td></tr>';
 	}
-	$text .= '</table><br />.:Добавить смайлик:.';
+	$text .= '</table>';
 	$sfiles = AdminSmilesGetAllSmiles($id, $smilies_dir, '', $smilies);
 	if(isset($sfiles[0])){
 		$fname = $sfiles[0]['name'];
 	}else{
 		$fname = '';
 	}
-	AddCenterBox('Смайлики');
 	AddText($text);
-	// ----------------------------------
 	if(count($sfiles) > 0){
-		FormRow('Изображение', $site->Select('file', $sfiles, false, "onchange=\"document.newsmile.image.src='$smilies_dir'+document.newsmile.file.value\""));
-		FormRow('Предпросмотр', "<img id=\"image\" src=\"$smilies_dir$fname\" />");
-		FormRow('Код', $site->Edit('code'));
-		FormRow('Описание', $site->Edit('desc'));
-		FormRow('Показывать', $site->Radio('indexview', 'on', true).'Да&nbsp;'.$site->Radio('indexview', 'off').'Нет');
-		AddForm("<form name=\"newsmile\" action=\"$mod&a=addsmile\" method=\"post\">", $site->Submit('Добавить'));
+		System::admin()->FormTitleRow('Добавить смайлик');
+		System::admin()->FormRow('Изображение', $site->Select('file', $sfiles, false, "onchange=\"document.newsmile.image.src='$smilies_dir'+document.newsmile.file.value\""));
+		System::admin()->FormRow('Предпросмотр', "<img id=\"image\" src=\"$smilies_dir$fname\" />");
+		System::admin()->FormRow('Код', $site->Edit('code', '', false, 'style="width: 200px;"'));
+		System::admin()->FormRow('Описание', $site->Edit('desc', '', false, 'style="width: 200px;"'));
+		System::admin()->FormRow('Показывать', $site->Radio('indexview', 'on', true).'Да&nbsp;'.$site->Radio('indexview', 'off').'Нет');
+		System::admin()->AddForm("<form name=\"newsmile\" action=\"$mod&a=addsmile\" method=\"post\">", $site->Submit('Добавить'));
 	}else{
-		AddText('<br />Новых файлов не найдено.');
+		System::admin()->Highlight('Загрузите изображения смайликов в папку <b>'.$smilies_dir.'</b> для добавления.');
 	}
 }
 
-function AdminSmilesEditSmile()
-{
+function AdminSmilesEditSmile(){
 	global $db, $config, $site, $smilies_dir, $mod;
 	$id = SafeEnv($_GET['id'], 11, int);
 	$smilies = $db->Select('smilies', '');
@@ -112,73 +145,59 @@ function AdminSmilesEditSmile()
 	$en[$smd['enabled']] = true;
 	FormRow('Изображение', $site->Select('file', $sfiles, false, "style=\"width:130px;\" onchange=\"document.newsmile.image.src='$smilies_dir'+document.newsmile.file.value\""));
 	FormRow('Предпросмотр', '<img id="image" src="'.$smilies_dir.$sfiles[$sid]['name'].'" />');
-	FormRow('Код', $site->Edit('code', $smd['code']));
-	FormRow('Описание', $site->Edit('desc', $smd['desc']));
+	FormRow('Код', $site->Edit('code', $smd['code'], false, 'style="width: 200px;"'));
+	FormRow('Описание', $site->Edit('desc', $smd['desc'], false, 'style="width: 200px;"'));
 	FormRow('Показывать', $site->Radio('indexview', 'on', $en[1]).'Да&nbsp;'.$site->Radio('indexview', 'off', $en[0]).'Нет');
 	AddCenterBox('Редактирование смайлика');
 	AddForm('<form name="newsmile" action="'.$mod.'&a=seditsave&id='.$id.'" method="post">', $site->Button('Отмена', 'onclick="history.go(-1)"').$site->Submit('Сохранить'));
 }
 
-function AdminSmilesEditSave()
-{
-	global $db, $config, $mod;
+function AdminSmilesEditSave(){
+	global $mod;
 	$id = SafeEnv($_GET['id'], 11, int);
 	$disp = EnToInt(SafeEnv($_POST['indexview'], 3, str));
 	$vals = Values('', SafeEnv($_POST['code'], 30, str), SafeEnv($_POST['desc'], 255, str), SafeEnv($_POST['file'], 255, str), $disp);
-	$db->Update('smilies', $vals, "`id`='$id'", true);
+	System::database()->Update('smilies', $vals, "`id`='$id'", true);
 	GO($mod);
 }
 
-function AdminSmilesAddSave()
-{
-	global $db, $config, $mod;
+function AdminSmilesAddSave(){
+	global $mod;
 	$disp = EnToInt(SafeEnv($_POST['indexview'], 3, str));
 	$vals = Values('', SafeEnv($_POST['code'], 30, str), SafeEnv($_POST['desc'], 255, str), SafeEnv($_POST['file'], 255, str), $disp);
-	$db->Insert('smilies', $vals);
+	System::database()->Insert('smilies', $vals);
 	GO($mod);
 }
 
-function AdminSmilesDeleteSmile()
-{
-	global $config, $db, $smilies_dir, $mod;
-	if(isset($_GET['ok']) && $_GET['ok'] == '1'){
-		$db->Delete('smilies', "`id`='".SafeEnv($_GET['sid'], 11, int)."'");
-		GO($mod);
-	}else{
-		$r = $db->Select('smilies', "`id`='".SafeEnv($_GET['sid'], 11, int)."'");
-		$text = 'Вы действительно хотите удалить смайлик <img src="'.$smilies_dir.$r[0]['file'].'" /> ?<br />'
-		."<a href=\"$mod&a=delsmile&sid=".SafeEnv($_GET['sid'], 11, int).'&ok=1">Да</a>'
-		.' &nbsp;&nbsp;&nbsp; <a href="javascript:history.go(-1)">Нет</a>
-		<br />
-		<br />';
-		AddTextBox("Внимание", $text);
-	}
+function AdminSmilesDeleteSmile(){
+	global $smilies_dir, $mod;
+	System::database()->Delete('smilies', "`id`='".SafeEnv($_GET['sid'], 11, int)."'");
+	GO($mod);
 }
 
-function AdminBlocksChangeStatus()
-{
-	global $config, $db, $mod;
-	$db->Select('smilies', "`id`='".SafeEnv($_GET['id'], 11, int)."'");
-	if($db->NumRows() > 0){
-		$r = $db->FetchRow();
+function AdminBlocksChangeStatus(){
+	global $mod;
+	System::database()->Select('smilies', "`id`='".SafeEnv($_GET['id'], 11, int)."'");
+	if(System::database()->NumRows() > 0){
+		$r = System::database()->FetchRow();
 		if(SafeDB($r['enabled'], 1, int) == 1){
 			$en = '0';
 		}else{
 			$en = '1';
 		}
-		$db->Update('smilies', "enabled='$en'", "`id`='".SafeEnv($_GET['id'], 11, int)."'");
+		System::database()->Update('smilies', "enabled='$en'", "`id`='".SafeEnv($_GET['id'], 11, int)."'");
 	}
 	GO($mod);
 }
 
-function AdminSmiliesAutoAdd()
-{
-	global $mod, $smilies_dir, $db, $site, $config;
+function AdminSmiliesAutoAdd(){
+	global $mod, $smilies_dir, $site;
+	AddCenterBox('Авто-добавление смайликов');
 
-	$smilies = $db->Select('smilies', '');
+	$smilies = System::database()->Select('smilies', '');
 	$sfiles = AdminSmilesGetAllSmiles($id, $smilies_dir, '', $smilies);
 	if(count($sfiles) == 0){
-		AddTextBox('Ошибка' ,'<br />Новых файлов не найдено. Загрузите изображения смайликов в папку: <b>'.$smilies_dir.'</b>.<br /><br />');
+		System::admin()->Highlight('Новых файлов не найдено. Загрузите изображения смайликов в папку: <b>'.$smilies_dir.'</b>.');
 		return;
 	}
 
@@ -192,9 +211,11 @@ function AdminSmiliesAutoAdd()
 		<th>Описание</th>
 		<th>Имя файла</th>
 		<th>Виден на главной</th>
+		<th>Удалить</th>
 	</tr>';
 
 	foreach($sfiles as $sm){
+		$func = System::admin()->SpeedConfirm('Удалить файл', $mod.'&a=delfile&name='.$sm['name'], 'images/admin/delete.png', 'Удалить файл с сервера?');
 		$text .= '<tr>'
 		.'<td>'.$site->Check('smilies[]', $sm['name'], true).'</td>'
 		.'<td><img src="'.$smilies_dir.$sm['name'].'" /></td>'
@@ -202,18 +223,16 @@ function AdminSmiliesAutoAdd()
 		.'<td>'.$site->Edit('desc['.$sm['name'].']', '', false, 'style="width:160px;"').'</td>'
 		.'<td>'.$sm['name'].'</td>'
 		.'<td>'.$site->Check('en['.$sm['name'].']', '1', true).'</td>'
+		.'<td>'.$func.'</td>'
 		.'</tr>';
 	}
 	$text .= '</table><br />';
 	$text .= $site->Submit('Добавить').'<br /><br />';
 	$text .= $site->FormClose();
-
-	AddCenterBox('Авто-добавление смайликов');
 	AddText($text);
 }
 
-function AdminSmiliesAutoSave()
-{
+function AdminSmiliesAutoSave(){
 	global $db, $mod;
 	foreach($_POST['smilies'] as $file){
 		$file = RealPath2(SafeEnv($file, 255, str));
@@ -226,50 +245,8 @@ function AdminSmiliesAutoSave()
 	GO($mod);
 }
 
-if(isset($_GET['a'])){
-	$action = $_GET['a'];
-}else{
-	$action = 'main';
+function AdminSmiliesDeleteFile(){
+	global $smilies_dir, $mod;
+	unlink(RealPath2($smilies_dir.$_GET['name']));
+	GO($mod.'&a=auto');
 }
-
-function AdminSmiles()
-{
-	global $action;
-
-	TAddToolLink('Смайлики', 'main', 'smilies');
-	TAddToolLink('Авто-добавление', 'auto', 'smilies&a=auto');
-	TAddToolBox($action);
-	
-	switch($action){
-		case 'main':
-			AdminSmilesMain();
-			break;
-		case 'addsmile':
-			AdminSmilesAddSave();
-			break;
-		case 'delsmile':
-			AdminSmilesDeleteSmile();
-			break;
-		case 'editsmile':
-			AdminSmilesEditSmile();
-			break;
-		case 'seditsave':
-			AdminSmilesEditSave();
-			break;
-		case 'changestatus':
-			AdminBlocksChangeStatus();
-			break;
-		case 'auto':
-			AdminSmiliesAutoAdd();
-			break;
-		case 'autosave':
-			AdminSmiliesAutoSave();
-			break;
-		default:
-			AdminSmilesMain();
-	}
-}
-
-AdminSmiles();
-
-?>
